move deps too
