+++ .eslintignore
@@ -1,4 +1,4 @@
 node_modules/*
-ecc/deps/*
+ecc/scripts/deps/*
 ecc/scripts/fallback.js
 web-test-runner.config.js
