+++ ecc/blocks/form-handler/controllers/terms-conditions-component-controller.js
@@ -1,6 +1,6 @@
 /* eslint-disable no-unused-vars */
 import { LIBS } from '../../../scripts/scripts.js';
-import HtmlSanitizer from '../../../deps/html-sanitizer.js';
+import HtmlSanitizer from '../../../scripts/deps/html-sanitizer.js';
 import { fetchThrottledMemoized } from '../../../scripts/utils.js';
 
 const { customFetch } = await import(`${LIBS}/utils/helpers.js`);
