+++ ecc/scripts/scripts.js
@@ -107,7 +107,7 @@ export const LIBS = (() => {
   return branch.includes('--') ? `https://${branch}.hlx.live/libs` : `https://${branch}--milo--adobecom.hlx.live/libs`;
 })();
 
-export const BlockMediator = await import('../deps/block-mediator.min.js').then((mod) => mod.default);
+export const BlockMediator = await import('./deps/block-mediator.min.js').then((mod) => mod.default);
 
 (function loadStyles() {
   const paths = [`${LIBS}/styles/styles.css`];
